# FrontEnd/src/openApi/openApi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/devisRequest/update": {
      "put": {
        "tags": [
          "devis-request-controller"
        ],
        "operationId": "updateDevisRequest",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DevisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/devis/update": {
      "put": {
        "tags": [
          "devis-controller"
        ],
        "operationId": "updateDevis",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Devis"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/customer/update": {
      "put": {
        "tags": [
          "customer-controller"
        ],
        "operationId": "updateCustomer",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Customer"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/cable/update": {
      "put": {
        "tags": [
          "cable-controller"
        ],
        "operationId": "updateCable",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Cable"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/battery/update": {
      "put": {
        "tags": [
          "battery-controller"
        ],
        "operationId": "updateBattery",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Battery"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/activity/update": {
      "put": {
        "tags": [
          "activity-controller"
        ],
        "operationId": "modifier_EvenementChantier",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Activity"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/devisRequest/add": {
      "post": {
        "tags": [
          "devis-request-controller"
        ],
        "operationId": "addDevisRequest",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DevisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/devis/add": {
      "post": {
        "tags": [
          "devis-controller"
        ],
        "operationId": "addDevis",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Devis"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/customer/add": {
      "post": {
        "tags": [
          "customer-controller"
        ],
        "operationId": "addCustomer",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Customer"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              }
            }
          }
        }
      }
    },
...
    "/cable/add": {
      "post": {
        "tags": [
          "cable-controller"
        ],
        "operationId": "addCable",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Cable"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/battery/add": {
      "post": {
        "tags": [
          "battery-controller"
        ],
        "operationId": "addBattery",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Battery"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/activity/add": {
      "post": {
        "tags": [
          "activity-controller"
        ],
        "operationId": "ajouter_jourTravailler",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Activity"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/devisRequest/get/{id}": {
      "get": {
        "tags": [
          "devis-request-controller"
        ],
        "operationId": "getDevisRequestById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DevisRequest"
                }
              }
            }
          }
        }
      }
    },
...
    "/devis/get/{id}": {
      "get": {
        "tags": [
          "devis-controller"
        ],
        "operationId": "getDevisById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Devis"
                }
              }
            }
          }
        }
      }
    },
...
    "/customer/get/{id}": {
      "get": {
        "tags": [
          "customer-controller"
        ],
        "operationId": "getCustomerById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              }
            }
          }
        }
      }
    },
...
    "/activity/get/{id}": {
      "get": {
        "tags": [
          "activity-controller"
        ],
        "operationId": "getActivitiesByWorkerId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Activity"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/devisRequest/delete/{id}": {
      "delete": {
        "tags": [
          "devis-request-controller"
        ],
        "operationId": "deleteDevisRequestById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/devis/delete/{id}": {
      "delete": {
        "tags": [
          "devis-controller"
        ],
        "operationId": "deleteDevis",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/customer/delete/{id}": {
      "delete": {
        "tags": [
          "customer-controller"
        ],
        "operationId": "deleteCustomer",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/cable/delete/{id}": {
      "delete": {
        "tags": [
          "cable-controller"
        ],
        "operationId": "deleteCable",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/battery/delete/{id}": {
      "delete": {
        "tags": [
          "battery-controller"
        ],
        "operationId": "deleteBattery",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/activity/delete/{id}": {
      "delete": {
        "tags": [
          "activity-controller"
        ],
        "operationId": "deleteActivity",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Activity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "start_date": {
            "type": "string",
            "format": "date-time"
          },
          "end_date": {
            "type": "string",
            "format": "date-time"
          },
          "text": {
            "type": "string"
          },
          "construction": {
            "type": "integer",
            "format": "int64"
          },
          "vehicule": {
            "type": "integer",
            "format": "int64"
          },
          "spots": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Spot"
            }
          },
          "workers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Worker"
            }
          }
        }
      },
      "Person": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "age": {
            "type": "integer",
            "format": "int32"
          },
          "address": {
            "type": "string"
          }
        }
      },
      "Spot": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "Worker": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "profession": {
            "type": "string"
          },
          "image": {
            "type": "string"
          },
          "generaleInfo": {
            "$ref": "#/components/schemas/Person"
          },
          "activities": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Activity"
            }
          }
        }
      },
...
      "DevisRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "location": {
            "type": "string"
          },
          "post_code": {
            "type": "string"
          },
          "building_type": {
            "type": "string"
          },
          "available_area": {
            "type": "integer",
            "format": "int32"
          },
          "roof_type": {
            "type": "string"
          },
          "electricity_access": {
            "type": "boolean"
          },
          "consumption": {
            "type": "number",
            "format": "float"
          },
          "status": {
            "type": "string"
          },
          "devis": {
            "type": "integer",
            "format": "int64"
          },
          "potential_customer": {
            "$ref": "#/components/schemas/PotentialCustomer"
          }
        }
      },
      "PotentialCustomer": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          }
        }
      },
      "Devis": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "ref": {
            "type": "string"
          },
          "etat": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "solarPanel": {
            "$ref": "#/components/schemas/MaterielSolarPanel"
          },
          "inverter": {
            "$ref": "#/components/schemas/MaterielInverter"
          },
          "systemFixing": {
            "$ref": "#/components/schemas/MaterielSystemFixing"
          },
          "cable": {
            "$ref": "#/components/schemas/MaterielCable"
          },
          "battery": {
            "$ref": "#/components/schemas/MaterielBattery"
          },
          "meter": {
            "$ref": "#/components/schemas/MaterielMeter"
          },
          "devisRequest": {
            "type": "integer",
            "format": "int64"
          },
          "construction": {
            "type": "integer",
            "format": "int64"
          },
          "total": {
            "type": "number",
            "format": "float"
          }
        }
      },
      "MaterielBattery": {
        "type": "object",
        "properties": {
          "modelBattery": {
            "type": "string"
          },
          "quantityBattery": {
            "type": "integer",
            "format": "int32"
          },
          "priceBattery": {
            "type": "number",
            "format": "float"
          },
          "tvaBattery": {
            "type": "integer",
            "format": "int32"
          },
          "totalBattery": {
            "type": "number",
            "format": "float"
          }
        }
      },
      "MaterielCable": {
        "type": "object",
        "properties": {
          "modelCable": {
            "type": "string"
          },
          "quantityCable": {
            "type": "integer",
            "format": "int32"
          },
          "priceCable": {
            "type": "number",
            "format": "float"
          },
          "tvaCable": {
            "type": "integer",
            "format": "int32"
          },
          "totalCable": {
            "type": "number",
            "format": "float"
          }
        }
      },
      "MaterielInverter": {
        "type": "object",
        "properties": {
          "modelInverter": {
            "type": "string"
          },
          "quantityInverter": {
            "type": "integer",
            "format": "int32"
          },
          "priceInverter": {
            "type": "number",
            "format": "float"
          },
          "tvaInverter": {
            "type": "integer",
            "format": "int32"
          },
          "totalInverter": {
            "type": "number",
            "format": "float"
          }
        }
      },
      "MaterielMeter": {
        "type": "object",
        "properties": {
          "modelMeter": {
            "type": "string"
          },
          "quantityMeter": {
            "type": "integer",
            "format": "int32"
          },
          "priceMeter": {
            "type": "number",
            "format": "float"
          },
          "tvaMeter": {
            "type": "integer",
            "format": "int32"
          },
          "totalMeter": {
            "type": "number",
            "format": "float"
          }
        }
      },
      "MaterielSolarPanel": {
        "type": "object",
        "properties": {
          "modelSolarPanel": {
            "type": "string"
          },
          "quantitySolarPanel": {
            "type": "integer",
            "format": "int32"
          },
          "priceSolarPanel": {
            "type": "number",
            "format": "float"
          },
          "tvaSolarPanel": {
            "type": "integer",
            "format": "int32"
          },
          "totalSolarPanel": {
            "type": "number",
            "format": "float"
          }
        }
      },
      "MaterielSystemFixing": {
        "type": "object",
        "properties": {
          "modelSystemFixing": {
            "type": "string"
          },
          "quantitySystemFixing": {
            "type": "integer",
            "format": "int32"
          },
          "priceSystemFixing": {
            "type": "number",
            "format": "float"
          },
          "tvaSystemFixing": {
            "type": "integer",
            "format": "int32"
          },
          "totalSystemFixing": {
            "type": "number",
            "format": "float"
          }
        }
      },
...
      "Construction": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "description": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "bill": {
            "type": "integer",
            "format": "int64"
          },
          "activities": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Activity"
            }
          },
          "amendment": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Amendment"
            }
          },
          "devis": {
            "type": "integer",
            "format": "int64"
          },
          "customer": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "Customer": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "generaleInfo": {
            "$ref": "#/components/schemas/Person"
          },
          "constructions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Construction"
            }
          }
        }
      },
      "Cable": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "model": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "length": {
            "type": "number",
            "format": "float"
          },
          "diameter": {
            "type": "number",
            "format": "float"
          },
          "nominal_voltage": {
            "type": "number",
            "format": "float"
          },
          "section_transversal": {
            "type": "string"
          },
          "material": {
            "type": "string"
          },
          "isolation": {
            "type": "string"
          },
          "resistance": {
            "type": "number",
            "format": "float"
          },
          "price": {
            "type": "number",
            "format": "float"
          }
        }
      },
      "Battery": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "model": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "storage_capacity": {
            "type": "number",
            "format": "float"
          },
          "nominal_voltage": {
            "type": "number",
            "format": "float"
          },
          "maximum_load_voltage": {
            "type": "number",
            "format": "float"
          },
          "maximum_discharge_voltage": {
            "type": "number",
            "format": "float"
          },
          "life_cycle": {
            "type": "string"
          },
          "operating_temperature": {
            "type": "number",
            "format": "float"
          },
          "lifespan": {
            "type": "string"
          },
          "date_manufacture": {
            "type": "string",
            "format": "date-time"
          },
          "price": {
            "type": "number",
            "format": "float"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          },
          "image": {
            "type": "string"
          }
        }
      },
...
    }
  }
}